+++ src/pages/ShoppingCartPage/ShoppingCartPage.tsx
@@ -4,7 +4,7 @@ import { useLoaderData, Await, Link } from 'react-router-dom';
 import Header from '../../components/Header/Header';
 import TitleContainer from '../../components/Container/TitleContainer/TitleContainer';
 import SubmitButton from '../../components/Button/SubmitButton/SubmitButton';
-import CartItemList from '../../components/CartItemList/CartItemList';
+import CartItemList from '../../components/List/CartItemList/CartItemList';
 import TotalPriceContainer from '../../components/Container/TotalPriceContainer/TotalPriceContainer';
 import type { CartItem } from '../../types/CartItem.type';
 import { selectedCartItemListState } from '../../recoil/CartItem/atoms/selectedCartItemListState';
